backend/pyproject.toml
[project]
name = "managed-chatkit-backend"
version = "0.1.0"
description = "FastAPI backend for creating ChatKit workflow sessions"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.114,<0.116",
    "httpx>=0.27,<0.28",
    "google-api-python-client>=2.147,<3.0",
    "google-auth>=2.35,<3.0",
    "google-auth-httplib2>=0.2,<0.3",
    "google-auth-oauthlib>=1.2,<2.0",
    "python-dateutil>=2.9,<3.0",
    "uvicorn[standard]>=0.36,<0.37",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.4,<0.7",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
